Bookmarking in Firefox now a two-stage prosess
Hi, I have no idea if this is the right place to post this but I was wondering why bookmarking a page in Firefox is now a two stage process requiring you to 1) bookmark a page and then 2) sort that bookmark into a folder? Clearly the old system was too user-friendly but I think we can fix this further by going to a five stage process

1): User presses bookmark star.
2): User forced to solve numeric puzzle.
3) Bookmark is placed in a random folder (Guess where!)
4) User has to hunt out bookmark manually and sort it into the folder they actually want it in.
5) Mozilla then posts the bookmark information printed on an index card forcing users to file them in rotary indexes and manually type out every bookmarked URL they wish to go to. (Users no longer able to open bookmarks using the old-fashioned 'open bookmark' button.)

I feel this would represent a massive improvement in the product. I would also like to suggest that Firefox should keep on opening bookmarks right at the end of the row instead of directly next to the tab they have open, forcing them to drag it back across to keep their tabs organized. Working together we can find more and more ways to annoy the user until they eventually transfer to another browser.